src/pyproject.toml
[tool.poetry]
name = "yopsbackend"
version = "0.1.0"
description = ""
authors = ["ivan <ivaneev.ivane@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.1"
djoser = "^2.2.2"
django = "^5.0.3"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.9"
django-cors-headers = "^4.3.1"
yookassa = "^3.0.1"
celery = {extras = ["redis"], version = "^5.3.6"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
